test(ui): Move route component tests out of source files
As in the previous commit, move the pure helpers from
`repository-runs-table.tsx` and `run-configuration-diff-dialog.tsx` to
`*-utils.ts` files so the external tests can import them.

The reasoning for this is given in the previous commit.

This concludes moving of all currently existing component tests out of
source files.
