+++ ui/src/routes/organizations/$orgId/products/$productId/repositories/$repoId/runs/$runIndex/dependencies/-components/dependency-tree-node.tsx
@@ -18,7 +18,6 @@
  */
 
 import { ChevronRight } from 'lucide-react';
-import { renderToStaticMarkup } from 'react-dom/server';
 
 import type { DependencyGraph } from '@/api';
 import { Badge } from '@/components/ui/badge';
@@ -164,111 +163,3 @@ export const DependencyTreeNode = ({
     </div>
   );
 };
-
-//
-// Unit tests for dependency tree rendering.
-//
-
-if (import.meta.vitest) {
-  const { describe, expect, it } = import.meta.vitest;
-
-  const graph: DependencyGraph = {
-    edges: [
-      { from: 0, to: 1 },
-      { from: 1, to: 0 },
-    ],
-    nodes: [
-      { fragment: 0, linkage: 'PROJECT_DYNAMIC', packageCount: 1, pkg: 0 },
-      { fragment: 1, linkage: 'DYNAMIC', packageCount: 1, pkg: 1 },
-    ],
-    packageCount: 2,
-    packages: [
-      { name: 'root', namespace: 'com.example', type: 'Maven', version: '1.0' },
-      {
-        name: 'library',
-        namespace: 'com.example',
-        type: 'Maven',
-        version: '2.0',
-      },
-    ],
-    purls: [null as unknown as string, 'pkg:maven/com.example/library@2.0'],
-    projectGroups: [],
-  };
-  const adjacency = new Map<number, number[]>([
-    [0, [1]],
-    [1, [0]],
-  ]);
-
-  describe('DependencyTreeNode', () => {
-    it('renders the node label and linkage badge', () => {
-      const markup = renderToStaticMarkup(
-        <DependencyTreeNode
-          adjacency={adjacency}
-          graph={graph}
-          matchesNodeSubtree={() => true}
-          nodeIndex={0}
-          packageIdType='ORT_ID'
-          path={new Set<number>()}
-        />
-      );
-
-      expect(markup).toContain('Maven:com.example:root:1.0');
-      expect(markup).toContain('PROJECT_DYNAMIC');
-    });
-
-    it('renders a cycle badge for cyclic child nodes', () => {
-      const markup = renderToStaticMarkup(
-        <DependencyTreeNode
-          adjacency={adjacency}
-          graph={graph}
-          matchesNodeSubtree={() => true}
-          nodeIndex={1}
-          packageIdType='ORT_ID'
-          path={new Set<number>([0, 1])}
-          searchTerm='library'
-        />
-      );
-
-      expect(markup).toContain('cycle');
-    });
-
-    it('renders only matching branches when children are filtered', () => {
-      const markup = renderToStaticMarkup(
-        <DependencyTreeNode
-          adjacency={new Map<number, number[]>([[0, [1]]])}
-          graph={{
-            ...graph,
-            edges: [{ from: 0, to: 1 }],
-          }}
-          matchesNodeSubtree={(nodeIndex) => nodeIndex === 1}
-          nodeIndex={0}
-          packageIdType='ORT_ID'
-          path={new Set<number>()}
-          searchTerm='library'
-        />
-      );
-
-      expect(markup).toContain('Maven:com.example:root:1.0');
-      expect(markup).toContain('library');
-      expect(markup).toContain('DYNAMIC');
-    });
-
-    it('renders the purl when purls are preferred', () => {
-      const markup = renderToStaticMarkup(
-        <DependencyTreeNode
-          adjacency={new Map<number, number[]>([[0, [1]]])}
-          graph={{
-            ...graph,
-            edges: [{ from: 0, to: 1 }],
-          }}
-          matchesNodeSubtree={() => true}
-          nodeIndex={1}
-          packageIdType='PURL'
-          path={new Set<number>()}
-        />
-      );
-
-      expect(markup).toContain('pkg:maven/com.example/library@2.0');
-    });
-  });
-}
